# apps/api/docs/swagger.yaml
basePath: /
definitions:
...
  handlers.JSONResult:
    type: object
...
  model.ActivityRequest:
    properties:
      hours:
        type: integer
      id:
        type: integer
      name:
        type: string
    type: object
...
host: 188.72.77.222:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Реализованный в рамках хакатона сервер для корпоративной системы благотворительности
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: АПИ сервера для корпоративной системы благотворительности
  version: "1.0"
paths:
  /ActivityRequest/{id}:
    delete:
      consumes:
      - application/json
      description: Delete item
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: 400 StatusBadRequest
          schema:
            type: string
        "404":
          description: 404 not found
          schema:
            type: string
      summary: Delete item
  /activity:
    get:
      consumes:
      - application/json
      description: Get all items
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.JSONResult'
            - properties:
                activities:
                  items: {}
                  type: array
              type: object
      summary: Get all items
    post:
      consumes:
      - application/json
      description: Add item
      parameters:
      - description: data
        in: body
        name: role
        required: true
        schema: {}
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
      summary: Add item
  /activity/{id}:
    delete:
      consumes:
      - application/json
      description: Delete item
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: 400 StatusBadRequest
          schema:
            type: string
        "404":
          description: 404 not found
          schema:
            type: string
      summary: Delete item
    get:
      consumes:
      - application/json
      description: Get one item
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
        "400":
          description: 400 StatusBadRequest
          schema:
            type: string
        "404":
          description: 404 not found
          schema:
            type: string
      summary: Get one item
    put:
      consumes:
      - application/json
      description: Update item
      parameters:
      - description: data
        in: body
        name: role
        required: true
        schema: {}
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
        "400":
          description: 400 StatusBadRequest
          schema:
            type: string
        "404":
          description: 404 not found
          schema:
            type: string
      summary: Update item
  /activity_request:
    get:
      consumes:
      - application/json
      description: Get all items
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.JSONResult'
            - properties:
                activities:
                  items:
                    $ref: '#/definitions/model.ActivityRequest'
                  type: array
              type: object
      summary: Get all items
    post:
      consumes:
      - application/json
      description: Add item
      parameters:
      - description: data
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/model.ActivityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ActivityRequest'
      summary: Add item
  /activity_request/{id}:
    get:
      consumes:
      - application/json
      description: Get one item
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ActivityRequest'
        "400":
          description: 400 StatusBadRequest
          schema:
            type: string
        "404":
          description: 404 not found
          schema:
            type: string
      summary: Get one item
    put:
      consumes:
      - application/json
      description: Update item
      parameters:
      - description: data
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/model.ActivityRequest'
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ActivityRequest'
        "400":
          description: 400 StatusBadRequest
          schema:
            type: string
        "404":
          description: 404 not found
          schema:
            type: string
      summary: Update item
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
  OAuth2Application:
    flow: application
    scopes:
      admin: ' Grants read and write access to administrative information'
      write: ' Grants write access'
    tokenUrl: https://example.com/oauth/token
    type: oauth2
swagger: "2.0"